% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_functions.R
\name{deflate_2016}
\alias{deflate_2016}
\title{Deflation to 2016 dollars, based on OMB deflators}
\usage{
deflate_2016(data_to_deflate, money_var = "Amount", fy_var = "Fiscal.Year",
  deflator_file = "2016_deflators_actuals.csv")
}
\arguments{
\item{data_to_deflate}{A data frame}

\item{money_var}{The quoted name of the dollar-value variable}

\item{fy_var}{The quoted name of the fiscal year variable}

\item{deflator_file}{The file name of the deflators to use; must be a CSV
saved on GitHub in CSISdefense/hamre_lookups/master}
}
\value{
A data frame with the money_var deflated, otherwise identical
to the original data frame
}
\description{
Deflation to 2016 dollars, based on OMB deflators
}
\section{Warning}{
 *This function should be used in data processing only, not in
a live app*.  It reads an external file from GitHub, which will slow down
an app substantially if done repeatedly.
}

\examples{
RDTE_data <- deflate_2016(
data_to_deflate = RDTE_data,
money_var = "Millions",
fy_var = "fiscal_year")
}
