% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{filter_by}
\alias{filter_by}
\title{A convenience function for filtering using quoted variable names}
\usage{
filter_by(data_frame, var_name, level_names, exclude = FALSE)
}
\arguments{
\item{data_frame}{The data frame}

\item{var_name}{The quoted name of the variable to use for filtering.}

\item{level_names}{The values (typically factor levels) of the variable
that you want to keep in the data, as vector of quoted value names.
Or, if you set exclude = TRUE, the variables you want to throw away.}

\item{exclude}{If TRUE, throw out rows with values listed in level_names.
If FALSE, keep rows with values listed in level_names.}
}
\value{
A filtered data frame
}
\description{
A convenience function for filtering using quoted variable names
}
\examples{
\dontrun{
shown \%<>\% filter_by(
  var_name = "Vendor.Size",
  level_names = input$vendor_size_filter)
}
}
