% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{hover_data}
\alias{hover_data}
\title{Retrieve the data for the current hover location}
\usage{
hover_data(chart_data, hover_object, chart_type = "line", max_distance = 10)
}
\arguments{
\item{chart_data}{A data frame.  The same data used to create the chart.}

\item{hover_object}{The Shiny input object containing the hover info.}

\item{chart_type}{Either "line" or "bar".  For scatterplot, use
"line", for area or stacked bar use "bar". Other types are not currently
supported and will require manual tips.}

\item{max_distance}{For line and scatter charts: a distance, in pixels.  If
the mouse is more than this distance from any point, the function will
return a null data frame.}
}
\value{
A data frame of
}
\description{
Retrieve the data for the current hover location
}
\examples{
\dontrun{
output$hover_info <- renderUI({

  hover_data(
    chart_data = dataset(),
    hover_object = input$plot_hover,
    group_by = c("Vendor.Size", "Simple"),
    chart_type = "stacked bar")
})
}
}
