% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{hover_style}
\alias{hover_style}
\title{Create a style for a tooltip wellPanel}
\usage{
wellPanel(style = hover_style(), "tip here")
}
\arguments{
\item{background_color}{A hex value, as quoted string, to form the
background color of the tooltip}

\item{preferred_side}{One of c("bottom_right", "top_right", "bottom_left",
"top_left")}

\item{flip_on_squish_h}{How close, in pixels, the mouse can get to the
(horizontal) edge of the chart before the tooltip flips to the other
side to give more space}

\item{flip_on_squish_v}{How close, in pixels, the mouse can get to the
(vertical) edge of the chart before the tooltip flips to the other
side to give more space}
}
\value{
A formatted string to include for wellPanel's style argument
within an output renderUI function for hover tips
}
\description{
Create a style for a tooltip wellPanel
}
\examples{
output$hover_info <- renderUI({
  if(is.null(input$plot_hover)){return()}
  year <- round(input$plot_hover$x)
  wellPanel(
    style = hover_style(),
    p(HTML(paste("hovering over year:", year))))
})

}
