% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{hover_tip}
\alias{hover_tip}
\title{Create a tooltip for a ggplot2 graph in Shiny}
\usage{
hover_tip(chart_data, hover_object, html_content, x_var = "Fiscal.Year",
  y_var = "Amount", chart_type = "line", ...)
}
\arguments{
\item{chart_data}{A data frame.  The same data used to create the chart.}

\item{hover_object}{The Shiny input object containing the hover info.}

\item{html_content}{The content to display in the tooltip.}

\item{x_var}{The quoted name of the variable on the X axis. NEVER MIND
YOU CAN DO THIS WITH hover$mapping$x}

\item{y_var}{The quoted name of the variable on the Y axis. NEVER MIND
YOU CAN DO THIS WITH hover$mapping$x}

\item{chart_type}{One of c("line", "bar", "stacked bar", "area").
Other types are not currently supported and will require manual tips.}

\item{...}{Additional arguments to customize the look of the tooltip;
will be passed through to hover_style()}
}
\value{
A wellPanel-based on-hover tooltip.
}
\description{
Create a tooltip for a ggplot2 graph in Shiny
}
\examples{
\dontrun{
output$hover_info <- renderUI({
  hover_tip(
    html_content = p(html("hello world")),
    chart_data = dataset(),
    hover_object = input$plot_hover,
    chart_type = "stacked bar")
})
}
}
