% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{money_label}
\alias{money_label}
\title{Converts numbers into character strings formatted as currency values,
with reasonable abbreviations and significant digits.  Is vectorized and
can be used with ggplot's axis scaling functions - see examples.}
\usage{
scale_y_continuous(labels = money_label)
}
\arguments{
\item{money_values}{A numeric or integer value or vector of values, to
convert to currency abbreviations}

\item{cur}{A character or string to prepend to the labels.  Likely a
currency sign - "$" by default.  Use "" for no prepended string.}
}
\value{
a character vector of formatted labels
}
\description{
Converts numbers into character strings formatted as currency values,
with reasonable abbreviations and significant digits.  Is vectorized and
can be used with ggplot's axis scaling functions - see examples.
}
\examples{
ggplot(
 data = FPDS \%>\%
   group_by(Fiscal.Year, Vendor.Size) \%>\%
   summarize(Action.Obligation = sum(Action.Obligation)),
 aes(
   x = Fiscal.Year,
   y = Action.Obligation,
   color = Vendor.Size)) +
geom_line() +
scale_y_continuous(labels = money_label)

}
