% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{money_labels}
\alias{money_labels}
\title{Tell ggplot to display dollar-value abbreviations on axis labels}
\usage{
scale_y_continuous(labels = money_labels)
}
\value{
a label-formatting function to be used with
scale_y_continuous or scale_x_continuous
}
\description{
Tell ggplot to display dollar-value abbreviations on axis labels
}
\examples{
ggplot(
 data = FPDS \%>\%
   group_by(Fiscal.Year, Vendor.Size) \%>\%
   summarize(Action.Obligation = sum(Action.Obligation)),
 aes(
   x = Fiscal.Year,
   y = Action.Obligation,
   color = Vendor.Size)) +
geom_line() +
scale_y_continuous(labels = money_labels)

}
