% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{set_wd_here}
\alias{set_wd_here}
\title{Sets the working directory to the folder holding the script from which it
is called.  This is kludgey and will only work if:}
\usage{
set_wd_here()
}
\description{
1. You are using RStudio, and
2. You run it directly from a script instead of the console or a source file.
}
\details{
It is intended to go at the start of data processing files
}
\examples{
\dontrun{
# Data processing for FPDS 3.0
library(tidyverse)
library(csis360)
set_wd_here()
}
}
