% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{sum_to}
\alias{sum_to}
\title{A convenience function for summation using quoted variable names}
\usage{
sum_to(data_frame, group_by)
}
\arguments{
\item{data_frame}{The data frame}

\item{group_by}{The variables you want to keep as breakouts in the
aggregated data, as a character vector of variable names.  This should
include any variable you plan to use on an X-axis (e.g. Fiscal.Year).}

\item{y_vars}{The variables you want to aggregate, as a character vector
of variable names.  By default, all numeric or integer variables that are
not listed in the group_by argument.  Variables that are not listed in
either group_by or y_vars will be rolled up in the aggregation and absent
from the returned data frame.}
}
\value{
An aggregated data frame
}
\description{
A convenience function for summation using quoted variable names
}
\examples{
\dontrun{
shown \%<>\% sum_to(
  group_by = c(input$breakout_variable, input$facet_variable, "Fiscal.Year"),
  y_vars = input$y_variable)
}
}
