% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{hover_style}
\alias{hover_style}
\title{Define the tooltip visual style for a tooltip on a ggplot2 graph
in a Shiny app}
\usage{
wellPanel(style = hover_style(input$plot_hover), "tip here")
}
\arguments{
\item{hover_object}{The Shiny input object containing the hover info.}

\item{background_color}{A hex value, as quoted string, to form the
background color of the tooltip}

\item{alpha}{A number between 0 and 1, to set the opacity of the
tooltip background}

\item{preferred_side}{One of c("right", "left")}

\item{h_just}{The horizontal distance, in pixels, to adjust the tooltip}

\item{v_just}{The vertical distance, in pixels, to adjust the tooltop}

\item{minimum_h}{The minimum horizonal width, in pixels, of the tooltip
panel.  The panel will squish to be smaller as the cursor approaches
the edge of the plotting area, but will not squish smaller than minimum_h.}

\item{minimum_v}{The minimum vertical width, in pixels, of the tooltip
panel.  The panel will "squish" to be smaller as the cursor approaches
the edge of the plotting area, but will not squish smaller than minimum_v.}
}
\value{
A wellPanel-based on-hover tooltip
}
\description{
Define the tooltip visual style for a tooltip on a ggplot2 graph
in a Shiny app
}
\examples{


}
