% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{hover_tip}
\alias{hover_tip}
\title{Create a tooltip for a ggplot in a Shiny application}
\usage{
hover_tip(input$plot_hover, "<b> Hello World </b>")
}
\arguments{
\item{hover_object}{The Shiny input object containing the hover info.}

\item{content}{*Either* a single-row data frame from which to automatically
generate a tooltip, *or* a string with the full contents of a custom
tooltip, formatted for HTML.}

\item{background_color}{A hex value, as string, to form the
background color of the tooltip}

\item{alpha}{A number between 0 and 1, to set the opacity of the
tooltip background}

\item{preferred_side}{One of c("right", "left")}

\item{h_just}{The horizontal distance, in pixels, to move the tooltip
to the right.  Negative values will move the tooltip to the left.}

\item{v_just}{The vertical distance, in pixels, to move the tooltip
up.  Negative values will move the tooltip down.}

\item{minimum_h}{The minimum horizonal width, in pixels, of the tooltip
panel.  The panel will squish to be smaller as the cursor approaches
the edge of the plotting area, but will not squish smaller than minimum_h.}

\item{minimum_v}{The minimum vertical width, in pixels, of the tooltip
panel.  The panel will "squish" to be smaller as the cursor approaches
the bottom of the plotting area, but will not squish smaller than minimum_v.}
}
\value{
A wellPanel-based on-hover tooltip
}
\description{
Create a tooltip for a ggplot in a Shiny application
}
\examples{
 \dontrun{
output$hover_info <- renderUI({
    
  hover_tip(
    hover_object = input$plot_hover,
    content = paste0(
      "<b> Fiscal Year: </b>", shown$Fiscal.Year, "<br/>",
      "<b> Vendor Size: </b>", shown$Vendor.Size, "<br/>",
      "<b> Amount: </b>", money_label(shown$Action.Obligation)))
})

output$hover_info <- renderUI({
  hover_tip(
    hover_object = input$plot_hover,
    content = hover_data(
      chart_data = dataset(),
      hover_object = input$plot_hover,
      chart_type = "bar"))
})

}
}
